fix: better valid-tool checking
Now using platform-specific methods and using the extended
`Player#hasCorrectToolForDrops` method on NeoForge, which fires
their harvesting event. Just using the default method on Fabric.

Also cleaned up PlatformUtil and split off client-only methods
into ClientPlatformUtil.
