+++ common/src/main/java/dev/ftb/mods/ftbultimine/client/ClientPlatformUtil.java
@@ -3,26 +3,10 @@ package dev.ftb.mods.ftbultimine.client;
 import dev.architectury.injectables.annotations.ExpectPlatform;
 import dev.ftb.mods.ftblibrary.ui.input.Key;
 import net.minecraft.client.KeyMapping;
-import net.minecraft.world.item.ItemStack;
 
-public class PlatformUtil {
+public class ClientPlatformUtil {
     @ExpectPlatform
     public static boolean doesKeybindMatch(KeyMapping keyMapping, int keyCode, int scanCode, int modifiers) {
-        throw new AssertionError();
-    }
-
-    @ExpectPlatform
-    public static boolean canAxeStrip(ItemStack stack) {
-        throw new AssertionError();
-    }
-
-    @ExpectPlatform
-    public static boolean canTillSoil(ItemStack stack) {
-        throw new AssertionError();
-    }
-
-    @ExpectPlatform
-    public static boolean canFlattenPath(ItemStack stack) {
         throw new AssertionError();
     }
