+++ common/src/main/java/dev/ftb/mods/ftbultimine/rightclick/FarmlandConversion.java
@@ -2,8 +2,8 @@ package dev.ftb.mods.ftbultimine.rightclick;
 
 import dev.ftb.mods.ftbultimine.api.rightclick.RightClickHandler;
 import dev.ftb.mods.ftbultimine.api.shape.ShapeContext;
-import dev.ftb.mods.ftbultimine.client.PlatformUtil;
 import dev.ftb.mods.ftbultimine.config.FTBUltimineServerConfig;
+import dev.ftb.mods.ftbultimine.utils.PlatformUtil;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerPlayer;
