+++ fabric/src/main/java/dev/ftb/mods/ftbultimine/client/fabric/ClientPlatformUtilImpl.java
@@ -1,0 +1,10 @@
+package dev.ftb.mods.ftbultimine.client.fabric;
+
+import net.minecraft.client.KeyMapping;
+
+public class ClientPlatformUtilImpl {
+    public static boolean doesKeybindMatch(KeyMapping keyMapping, int keyCode, int scanCode, int modifiers) {
+        // TODO how can we handle key modifiers on Fabric?
+        return keyMapping.matches(keyCode, scanCode);
+    }
+}
