+++ fabric/src/main/java/dev/ftb/mods/ftbultimine/utils/fabric/PlatformUtilImpl.java
@@ -1,15 +1,16 @@
-package dev.ftb.mods.ftbultimine.client.fabric;
+package dev.ftb.mods.ftbultimine.utils.fabric;
 
-import net.minecraft.client.KeyMapping;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.AxeItem;
 import net.minecraft.world.item.HoeItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ShovelItem;
+import net.minecraft.world.level.block.state.BlockState;
 
 public class PlatformUtilImpl {
-    public static boolean doesKeybindMatch(KeyMapping keyMapping, int keyCode, int scanCode, int modifiers) {
-        // TODO how can we handle key modifiers on Fabric?
-        return keyMapping.matches(keyCode, scanCode);
+    public static boolean playerHasCorrectTool(Player player, BlockPos pos, BlockState state) {
+        return player.hasCorrectToolForDrops(state);
     }
 
     public static boolean canAxeStrip(ItemStack stack) {
